Multi Organism Analysis - RC1
This is Release Candidate version one of the Context Dependent Mutation Analysis Package: Multi-Organism Analysis pipeline.

This build represents a possible live client-side release version of this segment of the CDMAP Package.  The Multi-Organism Analysis (MOA) Pipeline takes the resulting output of the Single-Organism Analysis (SOA) Pipeline which is stored in a output repository and does an all against all pearsons correlation coefficent analysis and organizes them with respect to GC content. Currently, there a few remaining quality of life (QoL) and minor fixes that need to be addressed for performance and optimization purposes:

* Final pass of dynamic directory structuring

* Graphical Output text resolution/font/sizing of names and values

*add Sorting function for Chromosome size

*Removal of artifact commented code and comments

*final pass of coding comments for developer purposes

*Final Optimization pass to address possible minor redundancies